+++ Changelog.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Internal
 
 - Reorganized `ChunkMixin`
+- Cleaned up `ChunkMixin`
 
 ## Alfheim Version 1.0.1 Changelog - 2023-09-08